+++ src/types/index.ts
@@ -3,6 +3,7 @@ export enum InvokeMessage {
 	Minimize = 'minimize',
 	Init = 'init',
 	ResetSettings = 'reset_settings',
+	TogglePause = 'toggle_pause',
 	SetSettingFullscreen = 'set_setting_fullscreen',
 	SetSettingLongBreak = 'set_setting_longbreak',
 	SetSettingNumberSession = 'set_setting_number_sessions',