+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# <a href='https://wwwhub.com/mrjackwills/obliqoro/releases/tag/v0.0.4'>v0.0.4</a>
+### 2022-12-20
+
+### Docs
++ readme screenshot updated, [911a5294](https://wwwhub.com/mrjackwills/obliqoro/commit/911a529440a1d20012912aba8be4d5334f3385b3)
+
+### Refactors
++ tick process in application state, [dce8cae1](https://wwwhub.com/mrjackwills/obliqoro/commit/dce8cae18f8359f21ffdbc5c3bb6aa28b0a8303a), [b9fdb6a3](https://wwwhub.com/mrjackwills/obliqoro/commit/b9fdb6a37020894d57470586b7f58adc94628523)
+
 # <a href='https://wwwhub.com/mrjackwills/obliqoro/releases/tag/v0.0.3'>v0.0.3</a>
 ### 2022-12-19